third_party/WebKit/LayoutTests/media/video-controls-overflow-menu-mute-button.html

<!DOCTYPE html>
<title>Clicking on the overflow mute button mutes the video.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<script src="overflow-menu.js"></script>

<!--Padding ensures the overflow menu is visible for the tests. -->
<body style="padding-top: 200px; padding-left: 100px">
<video controls></video>
<script>
async_test(function(t) {
  // Set up video
  var video = document.querySelector("video");
  video.src = "content/test.ogv";
  video.setAttribute("width", "60");
  video.setAttribute("height", "100");

  video.onloadeddata = t.step_func_done(function() {
    var overflowList = getOverflowList(video);
    var overflowMenu = getOverflowMenuButton(video);

    // Video is unmuted to begin with.
    assert_false(video.muted);

    // Click on the overflow menu button
    var coords = elementCoordinates(overflowMenu);
    clickAtCoordinates(coords[0], coords[1]);

    // Click on mute button
    var coords = elementCoordinates(overflowList.children[OverflowMenuButtons.MUTE]);
    clickAtCoordinates(coords[0], coords[1]);
    assert_true(video.muted);
  });
});
</script>
</body>
